t1415: avoid using `main` as ref name
In preparation for a patch series that will change the fall-back for
`init.defaultBranch` to `main`, let's not use `main` as ref name in this
test script.

Since we already use the name "second" for a secondary worktree that is
created in this test, let's use the name "first" for those refs instead.

While at it, adjust the test titles where "repo" was used by mistake
instead of the term "worktree".

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>