tests: drop prereq `PREPARE_FOR_MAIN_BRANCH` where no longer needed
We introduced the `PREPARE_FOR_MAIN_BRANCH` prereq for the sole purpose
of allowing us to perform the non-trivial adjustments regarding the
`master` -> `main` rename before the automatable ones.

Now that the transition is almost complete, we can stop using it in most
instances. The only exception is t9902: at the time of writing, there
are other patches in flight that touch this test script, therefore the
transition to `main` is postponed to a later date.

This patch is the result of this command:

	sed -i 's/PREPARE_FOR_MAIN_BRANCH[ ,]//' t/t[0-9]*.sh &&
	git checkout HEAD -- t/t9902\*

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>