t9832,t2200: avoid using pipes in git commands
When the upstream of a pipe throws an error, the downstream still
executes normally. This happens because the exit code of the upstream
in a pipe is ignored. This behavior can make debugging very hard
incase a test fails. Also, pipes are prone to deadlocks. If the
upstream gets full, the commands downstream will never start.

Write out the output of the git command to a file, so as to test the
exit codes of both commands.

Commit c6f44e1da5 (t9813: avoid using pipes, 2017-01-04) noticed that
the exit code of upstream in the pipe is ignored, thus using pipes
should be avoided.

Signed-off-by: Amanda Shafack <shafack.likhene@gmail.com>