hooks: allow input from stdin for commit-related hooks
Let hooks receive user input if applicable.

Closing stdin originates in f5bbc3225 (Port git commit to C, 2007).
Looks like the original shell implementation did have stdin open.

Due to stdin being closed, hooks that require user input have to either
read the input directly from the console (which can't work when running
from GUI applications), or popup a GUI dialog (which is inconvenient
when running from the terminal).

This allows for example prompting the user to choose an issue in
prepare-commit-msg, and add "Fixes #123" to the commit message.

Another possible use-case is running sanity test on pre-commit, and
having a prompt like "This and that issue were found in your changes.
Are you sure you want to commit? [Y/N]".

It's important to note that the hook author should be aware that stdin
is not always applicable. For example, when running from IDE. This can
be checked by isatty on stdin. The hooks should handle cases of closed
input, and possibly fall-back to GUI input, or have sane defaults with
a message to the user on this case.

Allow stdin only for commit-related hooks. Some of the other hooks pass
their own input to the hook, so don't change them.

Note: If pre-commit reads from stdin, and git commit is executed with
-F - (read message from stdin), stdin cannot be passed to the hook,
since it will consume it before reaching the point where it is read for
the commit message.

Signed-off-by: Orgad Shaneh <orgads@gmail.com>