hooks: allow input from stdin for commit-related hooks
Let hooks receive user input if applicable.

Closing stdin originates in f5bbc3225 (Port git commit to C,
2007). Looks like the original shell implementation did have
stdin open.

This allows for example prompting the user to choose an issue
in prepare-commit-msg, and add "Fixes #123" to the commit message.

Another possible use-case is running sanity test on pre-commit,
and having a prompt like "This and that issue were found in your
changes. Are you sure you want to commit? [Y/N]".

Allow stdin only for commit-related hooks. Some of the other
hooks pass their own input to the hook, so don't change them.

Note: If pre-commit reads from stdin, and git commit is executed
with -F - (read message from stdin), the message is not read
correctly. This is fixed in the follow-up commit.

Signed-off-by: Orgad Shaneh <orgads@gmail.com>