config: add --fixed-value option, un-implemented
The 'git config' builtin takes a 'value_regex' parameter for several
actions. This can cause confusion when expecting exact value matches
instead of regex matches, especially when the input string contains glob
characters. While callers can escape the patterns themselves, it would
be more friendly to allow an argument to disable the pattern matching in
favor of an exact string match.

Add a new '--fixed-value' option that does not currently change the
behavior. The implementation will follow for each appropriate action.
For now, check and test that --fixed-value will abort the command when
included with an incompatible action or without a 'value_regex'
argument.

The name '--fixed-value' was chosen over something simpler like
'--fixed' because some commands allow regular expressions on the
key in addition to the value.

Signed-off-by: Derrick Stolee <dstolee@microsoft.com>